refactor: moved SelectAllCheckBox to `TableColumnLabels` component
Now all the table-columns are rendered inside of a table row by the `TableColumnLabels` component. In doing so the local SelectAllCheckbox component was moved from the `Table` component to the `TableColumnLabels` component.
